Fixed solution for the operator homework

+++ src/OperatorHomework.java
@@ -1,21 +1,24 @@
 public class OperatorHomework {
 
     public static void main(String[] args) {
-        int i = 9;
-        int o = 18;
-        int k = 5;
-        int s = 7;
-        double y = i + o / k + k - (-i) * (++s);
+
+        float i = 9.0f;
+        float o = 18.0f;
+        float k = 5.0f;
+        float s = 7.0f;
+
+        double y = i + (o / k) + k - (-i) * (++s);
+
         System.out.println(y);
 
 
         // 1) (-i) = -9
-        // 2) (++s) = 1 + s = 1 + 7 = 8
+        // 2) (++s) = 1 + 7 = 8
         // 3) (o / k) = 18 / 5 = 3.6
-        // 4) i + 3.6 = 9 + 3.6 = 12.6
-        // 5) 12.6 + 5 = 17.6
-        // 6) 17.6 + 9 = 26.6
-        // 7) 26.6 * 8 = 212.8
+        // 4) -(-i) * (++s) = 9 * 8 = 72
+        // 5) i + 3.6 = 9 + 3.6 = 12.6
+        // 6) 12.6 + 5 = 17.6
+        // 7) 17.6 + 72 = 89.6
